automatically backed up by learn

+++ index.js
@@ -37,7 +37,7 @@ function dec(n) {
 }
 
 function makeInt(a){
-  parseFloat(a, 10)
+  parseInt(a, 10)
   return parseInt(a, 10)
 }
